+++ train_sml/__init__.py
@@ -1,1 +1,2 @@
-from .train_loop import UNet,train_model_source_separation,test_model_source_separation
+from .train_loop import train_model_source_separation,test_model_source_separation
+from .model_arch import UNet
